Modified the classDiagram, added arqDiagram. Moved SongDAO to models to have logic with persistance
