+++ src/ui/GUI.java
@@ -5,7 +5,7 @@ import java.awt.*;
 import java.awt.event.*;
 import java.util.Arrays;
 import model.*;
-import data.SongDAO;
+
 import java.util.List;
 import controller.JPlayer;
